Added the four_bit_LCD and the Servo projects
Now have 4 bit LCD and servo projects that work individually. However,
they do not yet work in conjunction- both currently use the P1.2 pin for
output. For PWM, the servo code needs to use either P1.2 or P1.6, both
of which are currently in use by the 4 bit LCD code